Update projects/detail.js to retry loading libs
Occasionally, the CodeMirror lib/mode that is being dynamically loaded
in projects/detail.js will fail. Without digging too far into internals
of RequireJS and tracing down the exact error, this change introspects
on the state of the actual page after attempting the load. If the libs
were loaded properly, then the CodeMirror snippet should be present on
the page. `checkLoaded()` will try to call the `load()` function three
times in total, with a yet-to-be-determined action for a third failure.