Make dynamic lib loading more robust
Occasionally, dynamically loading the CodeMirror library for displaying
code snippets will fail. In lieu of investigating the error from
RequireJS and addressing the specific cause, this implements a naive
retry mechanism that will introspect on the state of the DOM after
loading is supposed to have finished. If the codemirror div contains no
children, then loading has failed and will be retried a maximum of two
times (three load attempts in total) before displaying an error message
with a link to view the snippet on GitHub. This is not ideal, since it
requires manually verifying links after updating the projects they point
to, but works for me for now.