Set up for Capybara, add projects feature spec
Includes a small change to app/views/projects/index.html.erb. The
Capybara Selenium webdriver could not 'see' an empty <div> if the
closing tag was on the same line as the opening tag. Either including a
space between the opening and closing tags, or putting the closing tag
on a new line will make the <div> 'visible' to the webdriver