Dynamically set theme from query param
Theme may be sent in query param, such as
/presentations/coffee?theme=simple
When a theme is not found, a helpful page is shown with all options for
themes and a direct link to a randomly sampled theme.

Custom accent colors are set dynamically based on the theme being
categorized as 'light' or 'dark' in the controller.